backend/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.23.2
sqlmodel==0.0.25
psycopg2-binary==2.9.10
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic[email]==2.11.1
PyJWT==2.10.1
